cpp02 ex00 done

+++ cpp00/ex00/megaphone.cpp
@@ -1,7 +1,6 @@
 #include <iostream>
 #include <string>
 
-
 int main(int ac, char **av)
 {
 	if (ac < 2)
@@ -14,7 +13,7 @@ int main(int ac, char **av)
 	while (i < ac)
 	{
 		std::string str = av[i];
-		for (int j = 0; j < str.size(); j++)
+		for (size_t j = 0; j < str.size(); j++)
 			std::cout << (char)std::toupper(str[j]);
 		++i;
 	}
